Fix overflow handling
Early versions of the Arduino IDE had the timer overflow in the middle
of the variable range. This has changed a long time ago. So we can now
detect if a timer has run out by simply comparing now - start with the
duration of the timer which will be correct even with overflows.